NET CRACKER PROJECT. Version: 0.0.73 Branch: master Date: 28.05.2017 Author: AlexMerlyan
Working: refactor delete dao methods with IUDAnswer and services
Changes: dao delete methods and methods of services where was IUDAnswer
Next: refactor all delete dao and services delete methods

Comment: fuck yeah!

+++ src/main/java/projectpackage/repository/maintenancedao/ComplimentaryDAOImpl.java
@@ -10,6 +10,8 @@ import projectpackage.model.maintenances.Maintenance;
 import projectpackage.repository.AbstractDAO;
 import projectpackage.repository.daoexceptions.ReferenceBreakException;
 import projectpackage.repository.daoexceptions.TransactionException;
+import projectpackage.repository.daoexceptions.WrongEntityIdException;
+import projectpackage.repository.daoexceptions.DeletedObjectNotExistsException;
 import projectpackage.repository.reacteav.exceptions.ResultEntityNullException;
 
 import java.util.List;
@@ -73,7 +75,15 @@ public class ComplimentaryDAOImpl extends AbstractDAO implements ComplimentaryDA
     }
 
     @Override
-    public void deleteComplimentary(int id) throws ReferenceBreakException {
+    public void deleteComplimentary(int id) throws ReferenceBreakException, WrongEntityIdException, DeletedObjectNotExistsException {
+        Complimentary complimentary = null;
+        try {
+            complimentary = getComplimentary(id);
+        } catch (ClassCastException e) {
+            throw new WrongEntityIdException(this, e.getMessage());
+        }
+        if (null == complimentary) throw new DeletedObjectNotExistsException(this);
+
         deleteSingleEntityById(id);
     }
 }
